PR comment fix
